4.22 complete

+++ part4/bloglist/controllers/blogs.js
@@ -1,13 +1,22 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-	const blogs = await Blog.find({})
+	const blogs = await Blog.find({}).populate('user', {
+		username: 1,
+		name: 1,
+		id: 1,
+	})
 	response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response, next) => {
 	const body = request.body
+
+	const user = request.user
 
 	if (!body.likes) {
 		body.likes = '0'
@@ -18,22 +27,26 @@ blogsRouter.post('/', async (request, response, next) => {
 		author: body.author,
 		url: body.url,
 		likes: body.likes,
+		user: user._id,
 	})
-	try {
-		const savedBlog = await blog.save()
-		response.status(201).json(savedBlog.toJSON())
-	} catch (exception) {
-		next(exception)
-	}
+
+	const savedBlog = await blog.save()
+	user.blogs = user.blogs.concat(savedBlog._id)
+	await user.save()
+
+	response.status(201).json(savedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-	try {
-		await Blog.findByIdAndDelete(request.params.id)
-		response.status(204).end()
-	} catch (exception) {
-		next(exception)
+	const user = request.user
+
+	const blog = await Blog.findById(request.params.id)
+	if (blog.user.toString() !== user._id.toString()) {
+		return response.status(401).json({ error: 'user does not match' })
 	}
+
+	await Blog.findByIdAndDelete(request.params.id)
+	response.status(204).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
